# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API untuk Challenge 5",
        "version": "1.0.0",
        "description": "Dokumentasi API untuk Challenge 5"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
        "/api/v1/users": {
            "post": {
                "summary": "Menambahkan user dan profil baru",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "profile": {
                                        "type": "object",
                                        "properties": {
                                            "identity_type": {
                                                "type": "string"
                                            },
                                            "identity_number": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User dan profil baru ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request tidak valid"
                    },
                    "500": {
                        "description": "Kesalahan pada server"
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            },
            "get": {
                "summary": "Menampilkan data user",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Data user berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan pada server"
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            }
        },
        "/api/v1/users/{userId}": {
            "get": {
                "summary": "Menampilkan data user dan profilnya berdasarkan ID",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "description": "Masukkan ID user"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data user dan profil berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan pada server"
                    }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            },
            "put": {
                "summary": "Memperbarui data user dan profil berdasarkan ID",
                "tags": ["Users"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "schema": { "type": "integer" },
                        "required": true,
                        "description": "Masukkan ID user yang akan diupdate"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "email": { "type": "string" },
                                    "password": { "type": "string" },
                                    "profile": {
                                        "type": "object",
                                        "properties": {
                                            "identity_type": { "type": "string" },
                                            "identity_number": { "type": "string" },
                                            "address": { "type": "string" }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data user dan profil berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/User" }
                            }
                        }
                    },
                    "404": { "description": "Data user tidak ditemukan"},
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            },
            "delete": {
                "summary": "Menghapus data user dan profil berdasarkan ID",
                "tags": ["Users"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "schema": { "type": "integer" },
                        "required": true,
                        "description": "Masukkan ID user yang akan dihapus"
                    }
                ],
                "responses": {
                    "200": { "description": "User berhasil dihapus" },
                    "404": { "description": "User tidak ditemukan" },
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            }
        },
        "/api/v1/accounts": {
            "post": {
                "summary": "Menambahkan akun baru",
                "tags": ["Accounts"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": { "type": "integer" },
                                    "bank_name": { "type": "string" },
                                    "bank_account_number": { "type": "integer" },
                                    "balance": { "type": "number" }
                                },
                                "required": ["user_id", "bank_name", "bank_account_number", "amount"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Akun berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Account" }
                            }
                        }
                    },
                    "400": { "description": "Request atau input tidak valid" },
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            },
            "get": {
                "summary": "Menampilkan data akun yang telah ditambahkan",
                "tags": ["Accounts"],
                "responses": {
                    "200": {
                        "description": "Data akun berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Account" }
                                }
                            }
                        }
                    },
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            }
        },
        "/api/v1/accounts/{accountsId}": {
            "get": {
                "summary": "Menampilkan data akun berdasarkan ID",
                "tags": ["Accounts"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "accountsId",
                        "schema": { "type": "integer" },
                        "required": true,
                        "description": "Masukkan ID akun"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data akun berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Account" }
                            }
                        }
                    },
                    "404": { "description": "Akun tidak ditemukan" },
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            }
        },
        "/api/v1/transactions": {
            "post": {
                "summary": "Mengirimkan uang dari satu akun ke akun lain",
                "tags": ["Transactions"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "source_account_id": { "type": "integer" },
                                    "destination_account_id": { "type": "integer" },
                                    "amount": { "type": "number" }
                                },
                                "required": ["source_account_id", "destination_account_id", "amount"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi berhasil",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Transaction" }
                            }
                        }
                    },
                    "400": { "description": "Request tidak valid" },
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            },
            "get": {
                "summary": "Menampilkan data transaksi",
                "tags": ["Transactions"],
                "responses": {
                    "200": {
                        "description": "Data transaksi berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Transaction" }
                                }
                            }
                        }
                    },
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            }
        },
        "/api/v1/transactions/{transaction}": {
            "get": {
                "summary": "Menampilkan data transaksi berdasarkan ID",
                "tags": ["Transactions"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "transaction",
                        "schema": { "type": "integer" },
                        "required": true,
                        "description": "Masukkan ID transaksi yang akan ditampilkan"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data transaksi berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": { "type": "integer" },
                                        "source_account_id": { "$ref": "#/components/schemas/Account" },
                                        "destination_account_id": { "$ref": "#/components/schemas/Account" },
                                        "amount": { "type": "number" }
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Transaksi tidak ditemukan" },
                    "500": { "description": "Kesalahan pada server" }
                },
                "security": [
                    {
                        "basicAuth": [
                            "read:iniApiKey",
                            "write:iniApiKey"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    }
                }
            },
            "Profile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "identity_type": {
                        "type": "string"
                    },
                    "identity_number": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                }
            },
            "Account": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "bank_name": {
                        "type": "string"
                    },
                    "bank_account_number": {
                        "type": "integer"
                    },
                    "balance": {
                        "type": "number"
                    }
                }
            },
            "Transaction": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "account_id": {
                        "type": "integer"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}